requirements_fastapi.txt
fastapi==0.112.0
openai==1.40.6
python-dotenv==1.0.1
uvicorn==0.30.6
